Add external scanner for triple quoted strings
Summary:
[EEP64](https://www.erlang.org/eeps/eep-0064) defines triple quoted strings.
One of the features is that you can start with 3 or more quotes, and the ending delimiter must have the same amount.

This allows the use of triple-quotes inside a more-than-triple-quoted string.

But confirming the same number of quotes as starting and ending delimiter cannot be done with a regex.

So introduce an external scanner to do this counting.

Reviewed By: robertoaloi

Differential Revision: D58863944

fbshipit-source-id: c0e896ff8ac75d73eedd355d88d88372b13c65bf