Fix corner-case parsing lists with pipe
Summary:
# Context
In the tree-sitter grammar, the `|` symbol, used both for types and as list-consing operator, gets parsed as a binary-operator. While lowering lists, we check if the last expression of the list has a pipe at its root, and take the RHS as the tail of the list in that case.

# Problem
In a case like `[A = B | C]`, the content of the expr in the list would get parsed instead as `=(A, |(B,C)`, so lowering would consider it a parse error.

# This diff
The precedence of EQ was arguably wrong: e.g. it was lower than that of *,+, so that by default `A = B + C` would get parsed as `(A = B) + C`, and there was a `prec:right(...)` hack to compensate for this.

Instead, we increase the precedence of EQ, leaving it just below that of PIPE, so that `A = B + C |  D` parses as `(A = (B + C)) | D`

Reviewed By: michalmuskala

Differential Revision: D63701345

fbshipit-source-id: add68cc8a7c5b85b52ea1e6f8fa6b9bd31809372