1/n: Add initial ssr support
Summary:
# Diff stack overview

This diff stack makes a start on implementing Structural Search and Replace for ELP, modelled on what is [in rust-analyzer](https://rust-analyzer.github.io/manual.html#structural-search-and-replace).

The initial goal is to provide a simple template matching facility to make it easier to write diagnostics and assists. A later stack can expand this to provide replacement too.

It brings in a matching syntax which is made up of normal Erlang code, with special placeholder variables marking where it should match.  The placeholder syntax is based on what is used inside the `?Q` macro in [merl](https://www.erlang.org/doc/apps/syntax_tools/merl.html), so a template

```
{foo, _Foo}
```

would match the second element of a two part tuple.

A subsequent diff stack will provide tighter integration of this matching into the HIR helpers, similar to `codemod_helpers::find_call_in_function`.

# This diff

This diff introduces a new top-level form in the Erlang grammar, to capture a SSR template.  It defines a template as

```
ssr MATCH_SPEC ==>> REPLACEMENT when GUARDS .
```
The `==>> REPLACEMENT` and `when GUARDS` are optional, and not processed in this initial stack.

It also introduces hir constructors for `SsrPlaceHolder`, for `Expr`, `Pat`, and `TypeExpr`, and makes them invisible during a fold.

Reviewed By: robertoaloi

Differential Revision: D68214764

fbshipit-source-id: 5a2a276113ac82524737b4734dd3cfd85a9de03e