Bump tree-sitter and prebuildify deps
Summary:
The CI issue on Windows was caused by a vulnerability. This was fixed in:

* [`prebuildify`](https://github.com/prebuild/node-gyp-build/issues/68) 6.0.1
* [`tree-sitter`](https://github.com/tree-sitter/node-tree-sitter/commit/4017cfeca430eebf5cba69e517116e15634e0d02) 0.22.2

Bump the dependencies to include the respective fixes.

I am not sure if the Tree Sitter version has to be kept in sync with the one in `Cargo.toml`, so please advise. Also, I run:
```
make
make deps
make gen
```
To propagate the bump, not sure if more is needed.

Reviewed By: alanz

Differential Revision: D69656300

fbshipit-source-id: f73d873b510b9532f44dd7ee88028576a75db8d1