fix boolean precedence
Summary:
The tree-sitter grammar does not properly reflect the erlang precedence rules for binary operations.

The key changes are that we initially misunderstood the priority allocations, so it is reversed, and most importantly `orelse` and `andalso` had the highest, rather than lowest priority.

This diff aligns it with the erlc compiler, for both precedence and associativity.

Reviewed By: TD5

Differential Revision: D70894827

fbshipit-source-id: ebcb3321448e89876806618d1611e931cb2d95de